+++ src/screens/Login/index.js
@@ -1,15 +1,28 @@
 import React from 'react';
-import { View, Text } from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
 
-import TextInput from '@components/TextInput';
+import TextInput from 'components/TextInput';
+
+import StyleCommon from 'theme/StyleCommon';
 
 function Login() {
   return (
-    <View>
-      <Text> SIGN IN </Text>
-      <TextInput value="test" />
+    <View style={[StyleCommon.container, StyleCommon.centerContent]}>
+      <Text style={styles.title}> SIGN IN </Text>
+      <TextInput style={[StyleCommon.input, styles.input]} />
+      <TextInput style={StyleCommon.input} />
     </View>
   );
 }
+
+const styles = StyleSheet.create({
+  title: {
+    fontWeight: 'bold',
+    marginBottom: 20,
+  },
+  input: {
+    marginBottom: 20,
+  },
+});
 
 export default Login;
