+++ src/theme/StyleCommon.js
@@ -1,5 +1,21 @@
-import {StyleSheet} from 'react-native';
+import { StyleSheet } from 'react-native';
 
-const StyleCommon = StyleSheet.create({});
+const StyleCommon = StyleSheet.create({
+  centerContent: {
+    height: '100%',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  container: {
+    paddingHorizontal: 15,
+  },
+  input: {
+    width: '100%',
+    borderWidth: 1,
+    borderColor: '#000',
+    paddingVertical: 5,
+    paddingHorizontal: 5,
+  },
+});
 
 export default StyleCommon;
